+++ src/main/kotlin/com/chillin/adobe/response/AdobeAuthResponse.kt
@@ -1,11 +1,11 @@
-package com.chillin.connect.response
+package com.chillin.adobe.response
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 
 @JsonNaming(SnakeCaseStrategy::class)
-data class PrintSettingsResponse(
-    val id: String,
-    val uploadUri: String,
-) {
-}
+data class AdobeAuthResponse(
+    val accessToken: String,
+    val tokenType: String,
+    val expiresIn: Long,
+)
