+++ src/main/kotlin/com/chillin/epson/config/EpsonConnectConfig.kt
@@ -1,6 +1,7 @@
-package com.chillin.connect
+package com.chillin.epson.config
 
 import com.chillin.http.HttpClient
+import okhttp3.FormBody
 import okhttp3.OkHttpClient
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
@@ -13,5 +14,15 @@ class EpsonConnectConfig {
     @Bean
     fun epsonConnectClient(): HttpClient {
         return HttpClient(OkHttpClient())
+    }
+
+    @Bean
+    fun basicHeader(epsonConnectProperties: EpsonConnectProperties): String {
+        return epsonConnectProperties.basicHeader()
+    }
+
+    @Bean
+    fun authenticationHeader(epsonConnectProperties: EpsonConnectProperties): FormBody {
+        return epsonConnectProperties.authenticationForm()
     }
 }
