+++ src/main/kotlin/com/chillin/epson/response/EpsonConnectAuthResponse.kt
@@ -1,11 +1,14 @@
-package com.chillin.adobe.response
+package com.chillin.epson.response
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 
 @JsonNaming(SnakeCaseStrategy::class)
-data class AdobeAuthenticationResponse(
-    val accessToken: String,
+data class EpsonConnectAuthResponse(
     val tokenType: String,
+    val accessToken: String,
     val expiresIn: Long,
+    val refreshToken: String,
+    val subjectType: String,
+    val subjectId: String
 )
