+++ src/main/kotlin/com/chillin/epson/response/PrintSettingsResponse.kt
@@ -1,0 +1,10 @@
+package com.chillin.epson.response
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
+import com.fasterxml.jackson.databind.annotation.JsonNaming
+
+@JsonNaming(SnakeCaseStrategy::class)
+data class PrintSettingsResponse(
+    val id: String,
+    val uploadUri: String,
+)
