+++ src/test/kotlin/com/chillin/epson/EpsonConnectTest.kt
@@ -1,6 +1,6 @@
-package com.chillin.connect
+package com.chillin.epson
 
-import com.chillin.connect.request.PrintSettingsRequest
+import com.chillin.epson.request.PrintSettingsRequest
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestResult
